Move hitbox calculation to at collision detection
